+++ app/src/main/java/com/dlvjkb/hueapplication/model/lightbulbs/LightBulbState.java
@@ -12,10 +12,10 @@ public class LightBulbState implements Serializable {
     private ArrayList<Double> xy;
     private int ct;
     private String alert;
-    public int sat;
+    public float sat;
     private String effect;
-    public int bri;
-    public int hue;
+    public float bri;
+    public float hue;
     private String colorMode;
     private boolean reachable;
     public boolean on;
